requirements.txt
django==4.2.11
djangorestframework==3.15.1
channels==4.1.0
daphne==4.1.2
Pillow==10.3.0
channels-redis==4.2.0
